requirements.txt


snowflake-sqlalchemy
langchain
llama-cpp-python
torch
python-dotenv
snowflake-connector-python[pandas]
huggingface_hub




